Adds Lady Mondegreen and the Eggcorns pictures
For my final project during my first semester of HTML/CSS, I decided to create a website for a fictional Medieval band: Lady Mondegreen and the Eggcorns. It was incredibly fun to build--I had just learned so many new skills and was excited to practice all of them in one place.

This site also features several of my first JavaScript elements:

"Site last visited" dates that auto-update at the bottom of several web pages
The "Shows" page is generated dynamically based on the season
The "Featured Eggcorns" page has two buttons which, when clicked, display Philippa and Lettice Godfrey's different answers to a question regarding their strangest warm-up location before a performance
The site also contains forms that are all validated by regex and which submit information to a server on the Madison College campus.

Since this project, I've grown and changed a lot as a programmer, and picked up many new skills along the way. While the themes of the applications I've built since this website have become much more unified, I still have a soft spot for fancy buttons.