Adds Mad Libs PHP Project Images
This was my first project for my PHP development class. It represents the first time I built an application that connected my front-end knowledge with my back-end knowledge. I've always loved Mad Libs, so this application was a real treat to create, and to play with once it was done!