Adds screenshots/walkthrough of my Mad Libs app
This was the first PHP application I built in school. I love to play Mad Libs, so I had a grand old time building this game. This application also represents the first time I wrote a program that linked my front-end and back-end learning together--until this point, I hadn't communicated with a database AND designed a user interface all in one place, so this application really was the first day of the rest of my life. It was a highly satisfying introduction to full-stack developing.