Adds Couch Potato Fitness app screenshots
This was my second of four PHP projects, and it was also the one that taught me the most. It was comprehensive, combining not only what I'd learned in PHP web development at that point in my semester, but also what I'd learned in my HTML/CSS web development class before that. This was the last PHP application I built without assistance from the Bootstrap framework, so all of the styling here was completely hand-coded by me.